+++ examples/asr/asr_hybrid_transducer_ctc/speech_to_text_hybrid_rnnt_ctc_bpe_prompt.py
@@ -63,10 +63,10 @@ https://docs.nvidia.com/deeplearning/nemo/user-guide/docs/en/main/asr/configs.ht
 
 """
 
+import lightning.pytorch as pl
 from omegaconf import OmegaConf
 
-import lightning.pytorch as pl
-from nemo.collections.asr.models import EncDecHybridRNNTCTCBPEModelTgtLangID
+from nemo.collections.asr.models import EncDecHybridRNNTCTCBPEModelWithPrompt
 from nemo.core.config import hydra_runner
 from nemo.utils import logging
 from nemo.utils.exp_manager import exp_manager
@@ -74,14 +74,15 @@ from nemo.utils.trainer_utils import resolve_trainer_cfg
 
 
 @hydra_runner(
-    config_path="../conf/conformer/hybrid_transducer_ctc/", config_name="conformer_hybrid_transducer_ctc_bpe"
+    config_path="../conf/fastconformer/hybrid_transducer_ctc/",
+    config_name="fastconformer_hybrid_transducer_ctc_bpe_prompt.yaml",
 )
 def main(cfg):
     logging.info(f'Hydra config: {OmegaConf.to_yaml(cfg)}')
 
     trainer = pl.Trainer(**resolve_trainer_cfg(cfg.trainer))
     exp_manager(trainer, cfg.get("exp_manager", None))
-    asr_model = EncDecHybridRNNTCTCBPEModelTgtLangID(cfg=cfg.model, trainer=trainer)
+    asr_model = EncDecHybridRNNTCTCBPEModelWithPrompt(cfg=cfg.model, trainer=trainer)
 
     # Initialize the weights of the model from another model, if provided via config
     asr_model.maybe_init_from_pretrained_checkpoint(cfg)
