+++ nemo/collections/asr/models/hybrid_rnnt_ctc_bpe_models_tgt_lang.py
@@ -12,691 +12,16 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import os
-from typing import Any, Dict, Optional
+"""
+Temporary compatibility file to bridge the old model with the new prompt-based model.
+This file is needed to load old checkpoints and convert them to the new format.
+"""
 
-import torch
-from omegaconf import DictConfig, ListConfig, OmegaConf, open_dict
+from omegaconf import OmegaConf, open_dict
 
-from nemo.collections.asr.data.audio_to_text_dali import DALIOutputs
-from nemo.collections.asr.data.audio_to_text_lhotse_target_language import LhotseSpeechToTextBpeDatasetTgtLangID
-from nemo.collections.asr.metrics.bleu import BLEU
-from nemo.collections.asr.metrics.wer import WER
-from nemo.collections.asr.models.hybrid_rnnt_ctc_models import EncDecHybridRNNTCTCModel
-from nemo.collections.asr.parts.mixins import ASRBPEMixin, ASRTranscriptionMixin, TranscribeConfig
-from nemo.collections.asr.parts.submodules.ctc_decoding import CTCBPEDecoding, CTCBPEDecodingConfig
-from nemo.collections.asr.parts.submodules.rnnt_decoding import RNNTBPEDecoding
-from nemo.collections.common.data.lhotse import get_lhotse_dataloader_from_config
-from nemo.core.classes.common import typecheck
-from nemo.core.classes.mixins import AccessMixin
-from nemo.core.neural_types import (
-    AcousticEncodedRepresentation,
-    AudioSignal,
-    LabelsType,
-    LengthsType,
-    NeuralType,
-    SpectrogramType,
-)
-from nemo.utils import logging, model_utils
-from pytorch_lightning import Trainer
+from nemo.collections.asr.models.hybrid_rnnt_ctc_bpe_models_prompt import EncDecHybridRNNTCTCBPEModelWithPrompt
 
-
-class EncDecHybridRNNTCTCBPEModelTgtLangID(EncDecHybridRNNTCTCModel, ASRBPEMixin, ASRTranscriptionMixin):
-    """Base class for encoder decoder RNNT-based models with auxiliary CTC decoder/loss and subword tokenization."""
-
-    def __init__(self, cfg: DictConfig, trainer: Trainer = None):
-        # Convert to Hydra 1.0 compatible DictConfig
-        cfg = model_utils.convert_model_config_to_dict_config(cfg)
-        cfg = model_utils.maybe_update_config_version(cfg)
-
-        # Tokenizer is necessary for this model
-        if 'tokenizer' not in cfg:
-            raise ValueError("`cfg` must have `tokenizer` config to create a tokenizer !")
-
-        if not isinstance(cfg, DictConfig):
-            cfg = OmegaConf.create(cfg)
-
-        # Setup the tokenizer
-        self._setup_tokenizer(cfg.tokenizer)
-
-        # Initialize a dummy vocabulary
-        vocabulary = self.tokenizer.tokenizer.get_vocab()
-
-        # Set the new vocabulary
-        with open_dict(cfg):
-            cfg.labels = ListConfig(list(vocabulary))
-
-        with open_dict(cfg.decoder):
-            cfg.decoder.vocab_size = len(vocabulary)
-
-        with open_dict(cfg.joint):
-            cfg.joint.num_classes = len(vocabulary)
-            cfg.joint.vocabulary = ListConfig(list(vocabulary))
-            cfg.joint.jointnet.encoder_hidden = cfg.model_defaults.enc_hidden
-            cfg.joint.jointnet.pred_hidden = cfg.model_defaults.pred_hidden
-
-        # setup auxiliary CTC decoder
-        if 'aux_ctc' not in cfg:
-            raise ValueError(
-                "The config need to have a section for the CTC decoder named as aux_ctc for Hybrid models."
-            )
-
-        with open_dict(cfg):
-            if self.tokenizer_type == "agg":
-                cfg.aux_ctc.decoder.vocabulary = ListConfig(vocabulary)
-            else:
-                cfg.aux_ctc.decoder.vocabulary = ListConfig(list(vocabulary.keys()))
-
-        if cfg.aux_ctc.decoder["num_classes"] < 1:
-            logging.info(
-                "\nReplacing placholder number of classes ({}) with actual number of classes - {}".format(
-                    cfg.aux_ctc.decoder["num_classes"], len(vocabulary)
-                )
-            )
-            cfg.aux_ctc.decoder["num_classes"] = len(vocabulary)
-
-        # Setup language settings
-        with open_dict(cfg):
-            if not hasattr(cfg, 'language_settings'):
-                cfg.language_settings = OmegaConf.create({})
-            # Number of supported languages
-            if not hasattr(cfg.language_settings, 'num_languages'):
-                cfg.language_settings.num_languages = cfg.model_defaults.get('num_langs', 128)
-
-        super().__init__(cfg=cfg, trainer=trainer)
-
-        if cfg.get("initialize_target_lang_id_concatination", False):
-            self.initialize_target_lang_id_concatination()
-
-    def initialize_target_lang_id_concatination(self):
-        """Initialize model components for target language ID concatenation."""
-        logging.info("Parakeet model with target language support has been initalized")
-
-        # Setup decoding object
-        self.decoding = RNNTBPEDecoding(
-            decoding_cfg=self.cfg.decoding,
-            decoder=self.decoder,
-            joint=self.joint,
-            tokenizer=self.tokenizer,
-        )
-
-        # Setup wer object
-        self.wer = WER(
-            decoding=self.decoding,
-            batch_dim_index=0,
-            use_cer=self.cfg.get('use_cer', False),
-            log_prediction=self.cfg.get('log_prediction', True),
-            dist_sync_on_step=True,
-        )
-
-        # Setup bleu object
-        self.bleu = BLEU(decoding=self.decoding, tokenize=self.cfg.get('bleu_tokenizer', "13a"), log_prediction=True)
-
-        # Setup fused Joint step if flag is set
-        if self.joint.fuse_loss_wer:
-            self.joint.set_loss(self.loss)
-            self.joint.set_wer(self.wer)
-
-        # Setup CTC decoding
-        ctc_decoding_cfg = self.cfg.aux_ctc.get('decoding', None)
-        if ctc_decoding_cfg is None:
-            ctc_decoding_cfg = OmegaConf.structured(CTCBPEDecodingConfig)
-            with open_dict(self.cfg.aux_ctc):
-                self.cfg.aux_ctc.decoding = ctc_decoding_cfg
-        self.ctc_decoding = CTCBPEDecoding(self.cfg.aux_ctc.decoding, tokenizer=self.tokenizer)
-
-        # Setup CTC WER
-        self.ctc_wer = WER(
-            decoding=self.ctc_decoding,
-            use_cer=self.cfg.aux_ctc.get('use_cer', False),
-            dist_sync_on_step=True,
-            log_prediction=self.cfg.get("log_prediction", False),
-        )
-
-        # setting the RNNT decoder as the default one
-        self.cur_decoder = "rnnt"
-
-        if hasattr(self.cfg, 'initialize_target_lang_id_concatination'):
-            self.concat = True
-            self.num_langs = self.cfg.language_settings.num_languages
-            # Setup projection layers
-            proj_in_size = self.num_langs + self._cfg.model_defaults.enc_hidden
-            proj_out_size = self._cfg.model_defaults.enc_hidden
-
-            self.lang_kernel = torch.nn.Sequential(
-                torch.nn.Linear(proj_in_size, proj_out_size * 2),
-                torch.nn.ReLU(),
-                torch.nn.Linear(proj_out_size * 2, proj_out_size),
-            )
-
-    def _setup_dataloader_from_config(self, config: Optional[Dict]):
-        dataset = None
-        dataloader = None
-
-        # Only Lhotse dataset with language ID is supported
-        if config.get("use_lhotse") and "initialize_target_lang_id_concatination" in self.cfg:
-            # Create Lhotse dataset with language ID support
-            dataset = LhotseSpeechToTextBpeDatasetTgtLangID(tokenizer=self.tokenizer, cfg=config)
-
-            # Create dataloader from Lhotse dataset
-            dataloader = get_lhotse_dataloader_from_config(
-                config,
-                global_rank=self.global_rank,
-                world_size=self.world_size,
-                dataset=dataset,
-            )
-        else:
-            # Log warning for unsupported configuration
-            logging.warning(
-                "Language ID feature is required for this model. Only Lhotse dataset with LhotseSpeechToTextBpeDatasetTgtLangID is supported."
-            )
-
-        # Return the dataloader (will be None if configuration not supported)
-        return dataloader
-
-    def _setup_transcribe_dataloader(self, config: Dict) -> 'torch.utils.data.DataLoader':
-        """
-        Setup function for a temporary data loader which wraps the provided audio file.
-
-        Args:
-            config: A python dictionary which contains the following keys:
-            paths2audio_files: (a list) of paths to audio files. The files should be relatively short fragments. \
-                Recommended length per file is between 5 and 25 seconds.
-            batch_size: (int) batch size to use during inference. \
-                Bigger will result in better throughput performance but would use more memory.
-            temp_dir: (str) A temporary directory where the audio manifest is temporarily
-                stored.
-
-        Returns:
-            A pytorch DataLoader for the given audio file(s).
-        """
-        if 'manifest_filepath' in config:
-            manifest_filepath = config['manifest_filepath']
-            batch_size = config['batch_size']
-        else:
-            manifest_filepath = os.path.join(config['temp_dir'], 'manifest.json')
-            batch_size = min(config['batch_size'], len(config['paths2audio_files']))
-
-        dl_config = {
-            'manifest_filepath': manifest_filepath,
-            'sample_rate': self.preprocessor._sample_rate,
-            'labels': self.joint.vocabulary,
-            'batch_size': batch_size,
-            'trim_silence': False,
-            'shuffle': False,
-            'num_workers': config.get('num_workers', min(batch_size, os.cpu_count() - 1)),
-            'pin_memory': True,
-            'use_lhotse': True,
-            'use_bucketing': False,
-            'drop_last': False,
-            'lang_field': config.get('lang_field', 'target_lang'),
-            'initialize_target_lang_id_concatination': True,
-        }
-
-        if config.get("augmentor"):
-            dl_config['augmentor'] = config.get("augmentor")
-
-        temporary_datalayer = self._setup_dataloader_from_config(config=DictConfig(dl_config))
-        return temporary_datalayer
-
-    def _transcribe_forward(self, batch: Any, trcfg: TranscribeConfig):
-        encoded, encoded_len = self.forward(
-            input_signal=batch[0], input_signal_length=batch[1], target_lang_id=batch[4]
-        )
-
-        # Prepare output dictionary based on decoder type
-        if self.cur_decoder == "rnnt":
-            # RNNT Path - just use encoded outputs directly
-            output = dict(encoded=encoded, encoded_len=encoded_len)
-        else:
-            # CTC Path - compute logits from encoder output
-            logits = self.ctc_decoder(encoder_output=encoded)
-            output = dict(logits=logits, encoded_len=encoded_len)
-            del encoded
-
-        return output
-
-    @property
-    def input_types(self) -> Optional[Dict[str, NeuralType]]:
-        if hasattr(self.preprocessor, '_sample_rate'):
-            input_signal_eltype = AudioSignal(freq=self.preprocessor._sample_rate)
-        else:
-            input_signal_eltype = AudioSignal()
-
-        return {
-            "input_signal": NeuralType(('B', 'T'), input_signal_eltype, optional=True),
-            "input_signal_length": NeuralType(tuple('B'), LengthsType(), optional=True),
-            "processed_signal": NeuralType(('B', 'D', 'T'), SpectrogramType(), optional=True),
-            "processed_signal_length": NeuralType(tuple('B'), LengthsType(), optional=True),
-            "sample_id": NeuralType(tuple('B'), LengthsType(), optional=True),
-            "target_lang_id": NeuralType(('B', 'T', 'D'), LabelsType()),
-        }
-
-    @property
-    def output_types(self) -> Optional[Dict[str, NeuralType]]:
-        return {
-            "outputs": NeuralType(('B', 'D', 'T'), AcousticEncodedRepresentation()),
-            "encoded_lengths": NeuralType(tuple('B'), LengthsType()),
-        }
-
-    @typecheck()
-    def forward(
-        self,
-        input_signal=None,
-        input_signal_length=None,
-        processed_signal=None,
-        processed_signal_length=None,
-        target_lang_id=None,
-    ):
-        """
-        Forward pass of the model. Note that for RNNT Models, the forward pass of the model is a 3 step process,
-        and this method only performs the first step - forward of the acoustic model.
-
-        Please refer to the `training_step` in order to see the full `forward` step for training - which
-        performs the forward of the acoustic model, the prediction network and then the joint network.
-        Finally, it computes the loss and possibly compute the detokenized text via the `decoding` step.
-
-        Please refer to the `validation_step` in order to see the full `forward` step for inference - which
-        performs the forward of the acoustic model, the prediction network and then the joint network.
-        Finally, it computes the decoded tokens via the `decoding` step and possibly compute the batch metrics.
-
-        Args:
-            input_signal: Tensor that represents a batch of raw audio signals,
-                of shape [B, T]. T here represents timesteps, with 1 second of audio represented as
-                `self.sample_rate` number of floating point values.
-            input_signal_length: Vector of length B, that contains the individual lengths of the audio
-                sequences.
-            processed_signal: Tensor that represents a batch of processed audio signals,
-                of shape (B, D, T) that has undergone processing via some DALI preprocessor.
-            processed_signal_length: Vector of length B, that contains the individual lengths of the
-                processed audio sequences.
-            target_lang_id: Tensor that represents the target language ID embeddings,
-                of shape (B, T, D) where D is the number of supported languages.
-                Used for language-specific encoding via concatenation with acoustic features.
-
-        Returns:
-            A tuple of 2 elements -
-            1) The log probabilities tensor of shape [B, T, D].
-            2) The lengths of the acoustic sequence after propagation through the encoder, of shape [B].
-        """
-        has_input_signal = input_signal is not None and input_signal_length is not None
-        has_processed_signal = processed_signal is not None and processed_signal_length is not None
-        if (has_input_signal ^ has_processed_signal) is False:
-            raise ValueError(
-                f"{self} Arguments ``input_signal`` and ``input_signal_length`` are mutually exclusive "
-                " with ``processed_signal`` and ``processed_signal_len`` arguments."
-            )
-
-        if not has_processed_signal:
-            processed_signal, processed_signal_length = self.preprocessor(
-                input_signal=input_signal,
-                length=input_signal_length,
-            )
-
-        # Spec augment is not applied during evaluation/testing
-        if self.spec_augmentation is not None and self.training:
-            processed_signal = self.spec_augmentation(input_spec=processed_signal, length=processed_signal_length)
-
-        encoded, encoded_len = self.encoder(audio_signal=processed_signal, length=processed_signal_length)
-        encoded = torch.transpose(encoded, 1, 2)  # B * D * T -> B * T * D
-
-        if self.concat:
-            if target_lang_id.shape[1] > encoded.shape[1]:
-                target_lang_id = target_lang_id[:, : encoded.shape[1], :]
-
-            # Concatenate encoded states with language ID
-            concat_enc_states = torch.cat([encoded, target_lang_id], dim=-1)
-
-            # Apply joint projection
-            encoded = self.lang_kernel(concat_enc_states)
-
-        encoded = torch.transpose(encoded, 1, 2)  # B * T * D -> B * D * T
-        return encoded, encoded_len
-
-    def training_step(self, batch, batch_nb):
-        # Reset access registry
-        if AccessMixin.is_access_enabled():
-            AccessMixin.reset_registry(self)
-
-        if self.is_interctc_enabled():
-            AccessMixin.set_access_enabled(access_enabled=True)
-
-        signal, signal_len, transcript, transcript_len, target_lang_id = batch
-
-        # forward() only performs encoder forward
-        if isinstance(batch, DALIOutputs) and batch.has_processed_signal:
-            encoded, encoded_len = self.forward(processed_signal=signal, processed_signal_length=signal_len)
-        else:
-            encoded, encoded_len = self.forward(
-                input_signal=signal, input_signal_length=signal_len, target_lang_id=target_lang_id
-            )
-        del signal
-
-        # During training, loss must be computed, so decoder forward is necessary
-        decoder, target_length, states = self.decoder(targets=transcript, target_length=transcript_len)
-
-        if hasattr(self, '_trainer') and self._trainer is not None:
-            log_every_n_steps = self._trainer.log_every_n_steps
-            sample_id = self._trainer.global_step
-        else:
-            log_every_n_steps = 1
-            sample_id = batch_nb
-
-        if (sample_id + 1) % log_every_n_steps == 0:
-            compute_wer = True
-        else:
-            compute_wer = False
-
-        # If fused Joint-Loss-WER is not used
-        if not self.joint.fuse_loss_wer:
-            # Compute full joint and loss
-            joint = self.joint(encoder_outputs=encoded, decoder_outputs=decoder)
-            loss_value = self.loss(
-                log_probs=joint, targets=transcript, input_lengths=encoded_len, target_lengths=target_length
-            )
-
-            # Add auxiliary losses, if registered
-            loss_value = self.add_auxiliary_losses(loss_value)
-
-            # Reset access registry
-            # from RNNT training
-            # if AccessMixin.is_access_enabled():
-            #     AccessMixin.reset_registry(self)
-
-            tensorboard_logs = {
-                'learning_rate': self._optimizer.param_groups[0]['lr'],
-                'global_step': torch.tensor(self.trainer.global_step, dtype=torch.float32),
-            }
-
-            if compute_wer:
-                self.wer.update(
-                    predictions=encoded,
-                    predictions_lengths=encoded_len,
-                    targets=transcript,
-                    targets_lengths=transcript_len,
-                )
-                _, scores, words = self.wer.compute()
-                self.wer.reset()
-                tensorboard_logs.update({'training_batch_wer': scores.float() / words})
-
-        else:  # If fused Joint-Loss-WER is used
-            # Fused joint step
-            loss_value, wer, _, _ = self.joint(
-                encoder_outputs=encoded,
-                decoder_outputs=decoder,
-                encoder_lengths=encoded_len,
-                transcripts=transcript,
-                transcript_lengths=transcript_len,
-                compute_wer=compute_wer,
-            )
-
-            # Add auxiliary losses, if registered
-            loss_value = self.add_auxiliary_losses(loss_value)
-
-            # Reset access registry
-            # from RNNT training
-            # if AccessMixin.is_access_enabled():
-            #     AccessMixin.reset_registry(self)
-
-            tensorboard_logs = {
-                'learning_rate': self._optimizer.param_groups[0]['lr'],
-                'global_step': torch.tensor(self.trainer.global_step, dtype=torch.float32),
-            }
-
-            if compute_wer:
-                tensorboard_logs.update({'training_batch_wer': wer})
-
-        if self.ctc_loss_weight > 0:
-            log_probs = self.ctc_decoder(encoder_output=encoded)
-            ctc_loss = self.ctc_loss(
-                log_probs=log_probs, targets=transcript, input_lengths=encoded_len, target_lengths=transcript_len
-            )
-            tensorboard_logs['train_rnnt_loss'] = loss_value
-            tensorboard_logs['train_ctc_loss'] = ctc_loss
-            loss_value = (1 - self.ctc_loss_weight) * loss_value + self.ctc_loss_weight * ctc_loss
-            if compute_wer:
-                self.ctc_wer.update(
-                    predictions=log_probs,
-                    targets=transcript,
-                    targets_lengths=transcript_len,
-                    predictions_lengths=encoded_len,
-                )
-                ctc_wer, _, _ = self.ctc_wer.compute()
-                self.ctc_wer.reset()
-                tensorboard_logs.update({'training_batch_wer_ctc': ctc_wer})
-
-        # note that we want to apply interctc independent of whether main ctc
-        # loss is used or not (to allow rnnt + interctc training).
-        # assuming ``ctc_loss_weight=0.3`` and interctc is applied to a single
-        # layer with weight of ``0.1``, the total loss will be
-        # ``loss = 0.9 * (0.3 * ctc_loss + 0.7 * rnnt_loss) + 0.1 * interctc_loss``
-        loss_value, additional_logs = self.add_interctc_losses(
-            loss_value, transcript, transcript_len, compute_wer=compute_wer
-        )
-        tensorboard_logs.update(additional_logs)
-        tensorboard_logs['train_loss'] = loss_value
-        # Reset access registry
-        if AccessMixin.is_access_enabled():
-            AccessMixin.reset_registry(self)
-
-        # Log items
-        self.log_dict(tensorboard_logs)
-
-        return {'loss': loss_value}
-
-    def predict_step(self, batch, batch_idx, dataloader_idx=0):
-        # TODO: add support for CTC decoding
-        signal, signal_len, transcript, transcript_len, sample_id, target_lang_id = batch
-
-        # forward() only performs encoder forward
-        if isinstance(batch, DALIOutputs) and batch.has_processed_signal:
-            encoded, encoded_len = self.forward(processed_signal=signal, processed_signal_length=signal_len)
-        else:
-            encoded, encoded_len = self.forward(
-                input_signal=signal, input_signal_length=signal_len, target_lang_id=target_lang_id
-            )
-        del signal
-
-        best_hyp_text, all_hyp_text = self.decoding.rnnt_decoder_predictions_tensor(
-            encoder_output=encoded, encoded_lengths=encoded_len, return_hypotheses=False
-        )
-
-        sample_id = sample_id.cpu().detach().numpy()
-        return list(zip(sample_id, best_hyp_text))
-
-    def validation_pass(self, batch, batch_idx, dataloader_idx):
-        if self.is_interctc_enabled():
-            AccessMixin.set_access_enabled(access_enabled=True)
-
-        signal, signal_len, transcript, transcript_len, target_lang_id = batch
-
-        # forward() only performs encoder forward
-        if isinstance(batch, DALIOutputs) and batch.has_processed_signal:
-            encoded, encoded_len = self.forward(processed_signal=signal, processed_signal_length=signal_len)
-        else:
-            encoded, encoded_len = self.forward(
-                input_signal=signal, input_signal_length=signal_len, target_lang_id=target_lang_id
-            )
-        del signal
-
-        tensorboard_logs = {}
-        loss_value = None
-
-        # If experimental fused Joint-Loss-WER is not used
-        if not self.joint.fuse_loss_wer:
-            if self.compute_eval_loss:
-                decoder, target_length, states = self.decoder(targets=transcript, target_length=transcript_len)
-                joint = self.joint(encoder_outputs=encoded, decoder_outputs=decoder)
-
-                loss_value = self.loss(
-                    log_probs=joint, targets=transcript, input_lengths=encoded_len, target_lengths=target_length
-                )
-                tensorboard_logs['val_loss'] = loss_value
-
-            self.wer.update(
-                predictions=encoded,
-                predictions_lengths=encoded_len,
-                targets=transcript,
-                targets_lengths=transcript_len,
-            )
-            wer, wer_num, wer_denom = self.wer.compute()
-            self.wer.reset()
-
-            tensorboard_logs['val_wer_num'] = wer_num
-            tensorboard_logs['val_wer_denom'] = wer_denom
-            tensorboard_logs['val_wer'] = wer
-
-        else:
-            # If experimental fused Joint-Loss-WER is used
-            compute_wer = True
-
-            if self.compute_eval_loss:
-                decoded, target_len, states = self.decoder(targets=transcript, target_length=transcript_len)
-            else:
-                decoded = None
-                target_len = transcript_len
-
-            # Fused joint step
-            loss_value, wer, wer_num, wer_denom = self.joint(
-                encoder_outputs=encoded,
-                decoder_outputs=decoded,
-                encoder_lengths=encoded_len,
-                transcripts=transcript,
-                transcript_lengths=target_len,
-                compute_wer=compute_wer,
-            )
-            if loss_value is not None:
-                tensorboard_logs['val_loss'] = loss_value
-
-            tensorboard_logs['val_wer_num'] = wer_num
-            tensorboard_logs['val_wer_denom'] = wer_denom
-            tensorboard_logs['val_wer'] = wer
-
-        log_probs = self.ctc_decoder(encoder_output=encoded)
-        if self.compute_eval_loss:
-            ctc_loss = self.ctc_loss(
-                log_probs=log_probs, targets=transcript, input_lengths=encoded_len, target_lengths=transcript_len
-            )
-            tensorboard_logs['val_ctc_loss'] = ctc_loss
-            tensorboard_logs['val_rnnt_loss'] = loss_value
-            loss_value = (1 - self.ctc_loss_weight) * loss_value + self.ctc_loss_weight * ctc_loss
-            tensorboard_logs['val_loss'] = loss_value
-
-        # CTC WER calculation
-        self.ctc_wer.update(
-            predictions=log_probs,
-            targets=transcript,
-            targets_lengths=transcript_len,
-            predictions_lengths=encoded_len,
-        )
-        ctc_wer, ctc_wer_num, ctc_wer_denom = self.ctc_wer.compute()
-        self.ctc_wer.reset()
-        tensorboard_logs['val_wer_num_ctc'] = ctc_wer_num
-        tensorboard_logs['val_wer_denom_ctc'] = ctc_wer_denom
-        tensorboard_logs['val_wer_ctc'] = ctc_wer
-
-        # BLEU score calculation
-        self.bleu.update(
-            predictions=encoded, predictions_lengths=encoded_len, targets=transcript, targets_lengths=transcript_len
-        )
-        bleu_metrics = self.bleu.compute(return_all_metrics=True, prefix="val_")
-        tensorboard_logs.update(
-            {
-                'val_bleu_num': bleu_metrics['val_bleu_num'],
-                'val_bleu_denom': bleu_metrics['val_bleu_denom'],
-                'val_bleu_pred_len': bleu_metrics['val_bleu_pred_len'],
-                'val_bleu_target_len': bleu_metrics['val_bleu_target_len'],
-                'val_bleu': bleu_metrics['val_bleu'],
-            }
-        )
-        self.bleu.reset()
-
-        self.log('global_step', torch.tensor(self.trainer.global_step, dtype=torch.float32))
-
-        # Inter-CTC losses and additional logging
-        loss_value, additional_logs = self.add_interctc_losses(
-            loss_value,
-            transcript,
-            transcript_len,
-            compute_wer=True,
-            compute_loss=self.compute_eval_loss,
-            log_wer_num_denom=True,
-            log_prefix="val_",
-        )
-        if self.compute_eval_loss:
-            # overriding total loss value. Note that the previous
-            # rnnt + ctc loss is available in metrics as "val_final_loss" now
-            tensorboard_logs['val_loss'] = loss_value
-        tensorboard_logs.update(additional_logs)
-
-        # Reset access registry
-        if AccessMixin.is_access_enabled():
-            AccessMixin.reset_registry(self)
-
-        return tensorboard_logs
-
-    def validation_step(self, batch, batch_idx, dataloader_idx=0):
-        tensorboard_logs = self.validation_pass(batch, batch_idx, dataloader_idx)
-        if type(self.trainer.val_dataloaders) == list and len(self.trainer.val_dataloaders) > 1:
-            self.validation_step_outputs[dataloader_idx].append(tensorboard_logs)
-        else:
-            self.validation_step_outputs.append(tensorboard_logs)
-
-        return tensorboard_logs
-
-    def test_step(self, batch, batch_idx, dataloader_idx=0):
-        logs = self.validation_pass(batch, batch_idx, dataloader_idx=dataloader_idx)
-        test_logs = {name.replace("val_", "test_"): value for name, value in logs.items()}
-        if type(self.trainer.test_dataloaders) == list and len(self.trainer.test_dataloaders) > 1:
-            self.test_step_outputs[dataloader_idx].append(test_logs)
-        else:
-            self.test_step_outputs.append(test_logs)
-        return test_logs
-
-    def multi_validation_epoch_end(self, outputs, dataloader_idx: int = 0):
-        # Calculate validation loss if required
-        if self.compute_eval_loss:
-            val_loss_mean = torch.stack([x['val_loss'] for x in outputs]).mean()
-            val_loss_log = {'val_loss': val_loss_mean}
-        else:
-            val_loss_log = {}
-
-        # Calculate WER
-        wer_num = torch.stack([x['val_wer_num'] for x in outputs]).sum()
-        wer_denom = torch.stack([x['val_wer_denom'] for x in outputs]).sum()
-        tensorboard_logs = {**val_loss_log, 'val_wer': wer_num.float() / wer_denom}
-
-        # Calculate CTC WER if applicable
-        if self.ctc_loss_weight > 0:
-            ctc_wer_num = torch.stack([x['val_wer_num_ctc'] for x in outputs]).sum()
-            ctc_wer_denom = torch.stack([x['val_wer_denom_ctc'] for x in outputs]).sum()
-            tensorboard_logs['val_wer_ctc'] = ctc_wer_num.float() / ctc_wer_denom
-
-        # Calculate BLEU score
-        bleu_num = torch.stack([x['val_bleu_num'] for x in outputs]).sum(dim=0)
-        bleu_denom = torch.stack([x['val_bleu_denom'] for x in outputs]).sum(dim=0)
-        bleu_pred_len = torch.stack([x['val_bleu_pred_len'] for x in outputs]).sum(dim=0)
-        bleu_target_len = torch.stack([x['val_bleu_target_len'] for x in outputs]).sum(dim=0)
-
-        val_bleu = self.bleu._compute_bleu(bleu_pred_len, bleu_target_len, bleu_num, bleu_denom)
-        tensorboard_logs['val_bleu'] = val_bleu
-
-        # Finalize and log metrics
-        metrics = {**val_loss_log, 'log': tensorboard_logs}
-        self.finalize_interctc_metrics(metrics, outputs, prefix="val_")
-
-        return metrics
-
-    @property
-    def bleu(self):
-        return self._bleu
-
-    @bleu.setter
-    def bleu(self, bleu):
-        self._bleu = bleu
-
-
-# Language to Index mapping
+# Global map of language codes to indices that existed in the old model
 GLOBAL_LANG_MAP = {
     'en-US': 0,
     'en-GB': 1,
@@ -775,3 +100,63 @@ GLOBAL_LANG_MAP = {
     'sm-WS': 98,
     'to-TO': 99,
 }
+
+
+# For backward compatibility
+class EncDecHybridRNNTCTCBPEModelTgtLangID(EncDecHybridRNNTCTCBPEModelWithPrompt):
+    """Compatibility class for loading old checkpoints"""
+
+    def __init__(self, cfg, trainer=None):
+        # Add prompt_dictionary to cfg if not exists
+        if cfg is not None:
+            if not hasattr(cfg, 'model_defaults'):
+                with open_dict(cfg):
+                    cfg.model_defaults = OmegaConf.create({})
+
+            if not hasattr(cfg.model_defaults, 'prompt_dictionary'):
+                with open_dict(cfg.model_defaults):
+                    cfg.model_defaults.prompt_dictionary = GLOBAL_LANG_MAP
+
+            # Handle missing model section
+            if not hasattr(cfg, 'model'):
+                with open_dict(cfg):
+                    cfg.model = OmegaConf.create({})
+                    cfg.model.subsampling_factor = 8
+
+        super().__init__(cfg, trainer)
+
+        # Save num_langs for backward compatibility
+        self.num_langs = self.num_prompts
+
+    def load_state_dict(self, state_dict, strict=True):
+        # Map the old keys to new keys
+        new_state_dict = {}
+        for key, value in state_dict.items():
+            if key.startswith("lang_kernal"):
+                new_key = key.replace("lang_kernal", "lang_kernel")
+                new_state_dict[new_key] = value
+            else:
+                new_state_dict[key] = value
+
+        # Call the parent's load_state_dict with the fixed dictionary
+        return super().load_state_dict(new_state_dict, strict)
+
+    # Add compatibility methods
+    def convert_to_prompt_model(self, output_path):
+        """
+        Convert this model to a proper prompt-based model and save to a new checkpoint.
+
+        Args:
+            output_path: Path to save the new model checkpoint
+
+        Returns:
+            Path to the new model checkpoint
+        """
+        # The model is already using the prompt-based implementation
+        # Just need to save it with the new class
+        self.save_to(output_path)
+        return output_path
+
+
+# Alias for backwards compatibility
+EncDecHybridRNNTCTCBPEModelWithTgtLang = EncDecHybridRNNTCTCBPEModelTgtLangID
