+++ nemo/collections/asr/parts/utils/transcribe_utils.py
@@ -42,8 +42,8 @@ def get_buffered_pred_feat_rnnt(
     batch_size: int,
     manifest: str = None,
     filepaths: List[list] = None,
-    accelerator: Optional[str] = 'cpu',
     target_lang_id: str = None,
+    accelerator: Optional[str] = 'cpu',
 ) -> List[rnnt_utils.Hypothesis]:
     """
     Moved from examples/asr/asr_chunked_inference/rnnt/speech_to_text_buffered_infer_rnnt.py
@@ -52,6 +52,7 @@ def get_buffered_pred_feat_rnnt(
     hyps = []
     refs = []
     lang_ids = []
+
     if filepaths and manifest:
         raise ValueError("Please select either filepaths or manifest")
     if filepaths is None and manifest is None:
@@ -60,6 +61,7 @@ def get_buffered_pred_feat_rnnt(
     if manifest:
         filepaths = []
         with open(manifest, "r", encoding='utf_8') as mfst_f:
+            print("Parsing manifest files...")
             for L in mfst_f:
                 L = L.strip()
                 if not L:
@@ -69,6 +71,7 @@ def get_buffered_pred_feat_rnnt(
                 filepaths.append(audio_file)
                 if 'text' in row:
                     refs.append(row['text'])
+
                 # Extract language from manifest
                 if 'target_lang' in row:
                     lang_ids.append(row['target_lang'])
@@ -80,31 +83,35 @@ def get_buffered_pred_feat_rnnt(
     else:
         # If filepaths are provided directly, use lang_id from config for all
         lang_ids = [target_lang_id] * len(filepaths)
-        lang_ids.append(target_lang_id)
-
+        logging.info(f"filepaths are provided directly and target_lang_id: {target_lang_id}")
     with torch.inference_mode():
         with torch.amp.autocast('cpu' if accelerator == 'cpu' else 'cuda'):
             batch = []
-            batch_lang_ids = []
+            batch_lang_ids = []  # Initialize here instead of clearing
             asr.sample_offset = 0
             for idx in tqdm(range(len(filepaths)), desc='Sample:', total=len(filepaths)):
-                batch.append((filepaths[idx]))
+                batch.append(filepaths[idx])
                 batch_lang_ids.append(lang_ids[idx])
+
                 if len(batch) == batch_size:
                     audio_files = [sample for sample in batch]
 
+                    # Reset ASR for new batch
                     asr.reset()
+
                     # Set the language ID if any valid language ID exists
                     if any(lid is not None for lid in batch_lang_ids):
                         # Find the first non-None language ID to use
                         lang_id = next((lid for lid in batch_lang_ids if lid is not None), None)
                         if lang_id is not None:
-                            asr.target_lang_id = lang_id
+                            asr.set_target_lang_id(lang_id)
+
                     asr.read_audio_file(audio_files, delay, model_stride_in_secs)
                     hyp_list = asr.transcribe(tokens_per_chunk, delay)
                     hyps.extend(hyp_list)
 
                     batch.clear()
+                    batch_lang_ids.clear()
                     asr.sample_offset += batch_size
 
             if len(batch) > 0:
@@ -112,12 +119,19 @@ def get_buffered_pred_feat_rnnt(
                 asr.frame_bufferer.batch_size = len(batch)
                 asr.reset()
 
+                # Set the language ID for the remaining batch
+                if any(lid is not None for lid in batch_lang_ids):
+                    lang_id = next((lid for lid in batch_lang_ids if lid is not None), None)
+                    if lang_id is not None:
+                        asr.set_target_lang_id(lang_id)
+
                 audio_files = [sample for sample in batch]
                 asr.read_audio_file(audio_files, delay, model_stride_in_secs)
                 hyp_list = asr.transcribe(tokens_per_chunk, delay)
                 hyps.extend(hyp_list)
 
                 batch.clear()
+                batch_lang_ids.clear()
                 asr.sample_offset += len(batch)
 
     if os.environ.get('DEBUG', '0') in ('1', 'y', 't'):
@@ -388,6 +402,7 @@ def restore_transcription_order(manifest_path: str, transcriptions: list) -> lis
     reordered = [None] * len(transcriptions)
     for new, old in enumerate(new2old):
         reordered[old] = transcriptions[new]
+
     if is_list:
         reordered = tuple(map(list, zip(*reordered)))
     return reordered
@@ -666,4 +681,3 @@ class TextProcessingConfig:
 
     # Whether to separate punctuation with the previouse word by space.
     separate_punctuation: bool = True
-
